Restructure, add containerization, docker-compose Restructure question-service under backend
Added Dockerfile, docker-compose for configuration of services,
containerizing question-services and mongodb, binding volume for
persistence of db data
