+++ backend/question-service/.dockerignore
@@ -1,2 +1,2 @@
 node_modules
-data
+npm-debug.log
