+++ backend/question-service/config/db.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+    url: 'mongodb://mongodb:27017/question_db'
+};
