+++ backend/question-service/controllers/question.controller.js
@@ -1,4 +1,4 @@
-const db = require('../model');
+const db = require('../models');
 const Question = db.questions;
 
 exports.create = (req, res) => {
