+++ docker-compose.yml
@@ -1,0 +1,29 @@
+version: "3"
+services:
+#  user:
+#    build: ./backend/user-service
+#    container_name: user-service
+#    restart: unless-stopped
+#    ports:
+#      - 8001:8001
+#    links:
+#      - mongodb
+
+  question:
+    build: ./backend/question-service
+    container_name: question-service
+    restart: unless-stopped
+    ports:
+      - 8002:8002
+    links:
+      - mongodb
+
+  mongodb:
+    image: mongodb/mongodb-community-server:6.0-ubi8
+    container_name: mongodb
+    volumes:
+      - type: bind
+        source: ./data
+        target: /data/db
+    ports:
+      - 27017:27017
