--- question-service/config/db.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-    url: 'mongodb://127.0.0.1:27017/question_db'
-};
